调整文档格式&编译警告 (#85)
* fix -Wunused-function

* 调整文档格式

+++ src/observer/sql/stmt/select_stmt.cpp
@@ -46,7 +46,7 @@ RC SelectStmt::create(Db *db, const Selects &select_sql, Stmt *&stmt)
   // collect tables in `from` statement
   std::vector<Table *> tables;
   std::unordered_map<std::string, Table *> table_map;
-  for (int i = 0; i < select_sql.relation_num; i++) {
+  for (size_t i = 0; i < select_sql.relation_num; i++) {
     const char *table_name = select_sql.relations[i];
     if (nullptr == table_name) {
       LOG_WARN("invalid argument. relation name is null. index=%d", i);
